Fixed UI Issues on the Login Activity, Login Fragment and Sign Up Fragment.
Removed the unnecessary Status Updater Service and its Broadcast Receivers, since there is no use to notify the Server that you are online when the only use of being online is to see the hangouts around you - which can be Achieved by the LocationDetectorService which locating your coordinates every 5 minutes only while the app is active and can be turned on/off by the user anytime.

Transferred the Retrofit requests to server to get hangouts around, to get the list of all your hangouts and to get a specific hangout from the LocationDetectorService to the Main Activity, HangoutsAroundFragment, MyHangoutsFragment, Hangout Activity and ChatFragment, since those Retrofit requests to server are only relevant when the user is in one of these Activities/Fragments and there is no need to call them from an external service and pass the results through a Bundle inside an Intent - in this way the Tripal is much more stable and efficient in its performance.