Fixed UI Issues on the HangoutsAroundFragment (showing offline message on rotate, even when user is active), Main Activity (issues when the user rotates the screen) and MyHangoutsFragment (messages when there are no hangouts are not shown when a user leaves the hangout).
Minor changes in the Profile Activity Design (Transferred the change profile photo button inside the collapsing toolbar, changes the design of text).

Fixed issues with showing information on the Profile Activity/

Now profile photos will be shown on hangouts around, on active hangouts that the user participate in and on past hangouts.

Hangouts lists (hangouts around you and active hangouts that the user participate) will be updated when a user leaves the hangouts and when a user joins the hangout.